Move _can_save_longdouble and _can_serialize_meta to io
